fastapi_apps/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["bendeez <daniellacina8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.5"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.4.0"
aiomysql = "^0.2.0"
alembic = "^1.13.2"
cryptography = "^43.0.0"
pyjwt = "^2.9.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
passlib = "^1.7.4"
bcrypt = "^4.2.0"
slowapi = "^0.1.9"
httpx = "^0.27.2"
pytest-lazy-fixtures = "^1.1.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.243.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
